requirements.txt
streamlit
sentence-transformers>=2.2.0
faiss-cpu
groq
requests
urllib3
PyPDF2
python-docx
python-dotenv